*RDFTermFactoryTest -> *RDFTest
