+++ rdf4j/src/test/java/org/apache/commons/rdf/rdf4j/RDF4JTest.java
@@ -15,42 +15,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.rdf.jsonldjava;
+package org.apache.commons.rdf.rdf4j;
 
-import org.apache.commons.rdf.api.AbstractRDFTermFactoryTest;
+import org.apache.commons.rdf.api.AbstractRDFTest;
 import org.apache.commons.rdf.api.RDF;
 import org.junit.Assume;
-import org.junit.Ignore;
-import org.junit.Test;
 
-public class JsonLdFactoryTest extends AbstractRDFTermFactoryTest {
+public class RDF4JTest extends AbstractRDFTest {
 
 	@Override
 	public RDF createFactory() {
-		return new JsonLdRDF();
+		return new RDF4J();
 	}
 	
-	// TODO: Add support for checking for invalid lang/iri/blanknode IDs
-	
-	@Test
 	@Override
 	public void testInvalidLiteralLang() throws Exception {
-		Assume.assumeFalse("JSONLD-Java does not validate lang strings", false);		
+		Assume.assumeTrue("RDF4J doesn't check Lang strings",false);
 		super.testInvalidLiteralLang();
 	}
 	
-	@Test
-	@Override
-	public void testInvalidIRI() throws Exception {
-		Assume.assumeFalse("JSONLD-Java does not validate IRIs", false);
-		super.testInvalidIRI();
-	}
-	
-	@Ignore
-	@Test
-	@Override
-	public void testPossiblyInvalidBlankNode() throws Exception {
-		// TODO: Fix blank node in ntriplesString()
-		super.testPossiblyInvalidBlankNode();
-	}
 }
